refactor(api): extract the shared league week frame resolver
The Survivor dashboard glance owned the batched "which week is this
league on, and is it still open" resolution. Pick'em's glance needs the
identical answer, and two copies could disagree about which week a
member owes — so it moves to `resolveLeagueWeekFrames` in league-weeks,
beside the `resolveCurrentWeekId` it is built on. Survivor's behavior is
unchanged.
