+++ apps/api/src/services/survivor/pick-status.ts
@@ -1,6 +1,6 @@
-import { and, asc, inArray } from "drizzle-orm";
+import { and, inArray } from "drizzle-orm";
 import type { Db } from "@picksleagues/db";
-import { games, survivorPicks, survivorState, weeks } from "@picksleagues/db";
+import { survivorPicks, survivorState } from "@picksleagues/db";
 import type { Clock } from "@picksleagues/core";
 import {
   SURVIVOR_PICK_STATUS,
@@ -8,29 +8,19 @@ import {
   type SurvivorPickStatus,
   type SurvivorSettings,
 } from "@picksleagues/schemas";
-import { resolveGameOverrides } from "../games";
-import { resolveCurrentWeekId } from "../league-weeks";
-import { isLocked, isPickable } from "../slate";
+import { resolveLeagueWeekFrames } from "../league-weeks";
 import { isSurvivorRangeWeek, resolveSurvivorSeasonStates } from "./season";
 
 /**
  * The dashboard's one-line answer to "do I owe a pick here?" for every Survivor
  * league a member is in (spec §Screens — Dashboard).
  *
- * Batched across leagues rather than resolved per card: this sits on the
- * dashboard's critical path and a member can be in many leagues, so its query
- * count is constant whatever that count is — which is why the season-state
- * resolution below is the batched form too.
- *
- * Two rules the module exists to keep server-side:
- * - **Lock is derived, never stored** (arch D11) — computed from each game's
- *   effective kickoff against the injected Clock. Deliberately not computed in
- *   the browser: under the simulator the browser sits at a different instant
- *   from the API, so a browser-derived lock would contradict the very states
- *   every other surface computed.
- * - **The current week has one definition** — `resolveCurrentWeekId`, the same
- *   one the league week list answers with, so the dashboard cannot point a
- *   member at a different week than the pick screen does.
+ * Which week that is and whether it is still open comes from
+ * `resolveLeagueWeekFrames`, shared with Pick'em's glance — the derived-lock and
+ * one-definition-of-the-current-week rules it keeps are stated there. What is
+ * Survivor's own is everything below the frame: elimination and winning are
+ * facts about the member's *season*, so they outrank every week-shaped state
+ * and are answered before a week is even looked up.
  */
 
 /** A superset of `SurvivorSeasonStateInput`, so the batch below passes straight through. */
@@ -64,40 +54,16 @@ export async function resolveSurvivorPickStatuses(
 
   const leagueSeasonIds = leagues.map((league) => league.leagueSeasonId);
   const membershipIds = leagues.map((league) => league.membershipId);
-  const seasonIds = [...new Set(leagues.map((league) => league.seasonId))];
 
-  // Ordered by start instant like the week list, with the week number as a
-  // tiebreak so the "last played" fallback below lands on the same row twice.
-  const weekRows = await db
-    .select({
-      id: weeks.id,
-      seasonId: weeks.seasonId,
-      weekType: weeks.weekType,
-      weekNumber: weeks.weekNumber,
-      startsAt: weeks.startsAt,
-      endsAt: weeks.endsAt,
-    })
-    .from(weeks)
-    .where(inArray(weeks.seasonId, seasonIds))
-    .orderBy(asc(weeks.startsAt), asc(weeks.weekNumber));
-
-  const weeksBySeason = new Map<string, typeof weekRows>();
-  for (const row of weekRows) {
-    const bucket = weeksBySeason.get(row.seasonId);
-    if (bucket) bucket.push(row);
-    else weeksBySeason.set(row.seasonId, [row]);
-  }
-
-  // The range clip is per league, not per season: two leagues can share a
-  // season and configure different ranges over it.
-  const currentWeekByLeague = new Map<string, string>();
-  for (const league of leagues) {
-    const inRange = (weeksBySeason.get(league.seasonId) ?? []).filter((row) =>
-      isSurvivorRangeWeek(row, league.settings),
-    );
-    const weekId = resolveCurrentWeekId(inRange, clock);
-    if (weekId) currentWeekByLeague.set(league.leagueSeasonId, weekId);
-  }
+  const frames = await resolveLeagueWeekFrames(
+    db,
+    clock,
+    leagues.map((league) => ({
+      leagueSeasonId: league.leagueSeasonId,
+      seasonId: league.seasonId,
+      playsWeek: (week) => isSurvivorRangeWeek(week, league.settings),
+    })),
+  );
 
   const state = await db
     .select({
@@ -120,7 +86,7 @@ export async function resolveSurvivorPickStatuses(
       .map((row) => memberKey(row.leagueSeasonId, row.leagueMemberId)),
   );
 
-  const currentWeekIds = [...new Set(currentWeekByLeague.values())];
+  const currentWeekIds = [...new Set([...frames.values()].map((frame) => frame.weekId))];
 
   const picks =
     currentWeekIds.length === 0
@@ -142,22 +108,6 @@ export async function resolveSurvivorPickStatuses(
   const picked = new Set(
     picks.map((row) => `${memberKey(row.leagueSeasonId, row.leagueMemberId)}:${row.weekId}`),
   );
-
-  const gameRows =
-    currentWeekIds.length === 0
-      ? []
-      : await db.select().from(games).where(inArray(games.weekId, currentWeekIds));
-
-  const now = clock.now();
-  const weeksWithGames = new Set<string>();
-  const weeksStillOpen = new Set<string>();
-  for (const row of gameRows) {
-    weeksWithGames.add(row.weekId);
-    const effective = resolveGameOverrides(row);
-    if (isPickable(effective.status) && !isLocked(effective.kickoffAt, now)) {
-      weeksStillOpen.add(row.weekId);
-    }
-  }
 
   const seasonStates = await resolveSurvivorSeasonStates(db, leagues);
 
@@ -183,21 +133,17 @@ export async function resolveSurvivorPickStatuses(
       continue;
     }
 
-    const weekId = currentWeekByLeague.get(league.leagueSeasonId);
-    if (!weekId) continue;
+    const frame = frames.get(league.leagueSeasonId);
+    if (!frame) continue;
 
-    if (picked.has(`${key}:${weekId}`)) {
+    if (picked.has(`${key}:${frame.weekId}`)) {
       statuses.set(league.leagueSeasonId, SURVIVOR_PICK_STATUS.PICK_IN);
       continue;
     }
 
-    // A week whose schedule hasn't been ingested has closed against nobody —
-    // settlement's prefix refuses to grade a week holding no games (ADR-0025),
-    // so reporting a miss there would name an elimination that cannot happen.
-    const open = weeksStillOpen.has(weekId) || !weeksWithGames.has(weekId);
     statuses.set(
       league.leagueSeasonId,
-      open ? SURVIVOR_PICK_STATUS.PICK_NEEDED : SURVIVOR_PICK_STATUS.LOCKED,
+      frame.open ? SURVIVOR_PICK_STATUS.PICK_NEEDED : SURVIVOR_PICK_STATUS.LOCKED,
     );
   }
